LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-validity-valid-weekmonth-expected.txt


FAIL [INPUT in MONTH status] validity.valid must be false if validity.rangeOverflow is true assert_false: The validity.valid should be false. expected false got true
FAIL [INPUT in MONTH status] validity.valid must be false if validity.rangeUnderflow is true assert_false: The validity.valid should be false. expected false got true
FAIL [INPUT in MONTH status] validity.valid must be false if validity.stepMismatch is true assert_false: The validity.valid should be false. expected false got true
PASS [INPUT in MONTH status] validity.valid must be false if validity.valueMissing is true
FAIL [INPUT in WEEK status] validity.valid must be false if validity.rangeOverflow is true assert_false: The validity.valid should be false. expected false got true
FAIL [INPUT in WEEK status] validity.valid must be false if validity.rangeUnderflow is true assert_false: The validity.valid should be false. expected false got true
FAIL [INPUT in WEEK status] validity.valid must be false if validity.stepMismatch is true assert_false: The validity.valid should be false. expected false got true
PASS [INPUT in WEEK status] validity.valid must be false if validity.valueMissing is true
